+++ Assets/Packages/com.studio23.ss2.loadingsystem/Runtime/Core/SceneLoadingSystem.cs
@@ -31,39 +31,48 @@ namespace Studio23.SS2.SceneLoadingSystem.Core
         }
 
         /// <summary>
-        /// Load a single scene by scene name
+        /// Load a single sceneToUnload by sceneToUnload name
         /// </summary>
         /// <param name="scene"></param>
         /// <returns></returns>
-        public async UniTask LoadScene(string scene, bool makeSceneActive = false)
+        public async UniTask LoadScene(string scene)
         {
-            await LoadScenes(new List<string> { scene}, makeSceneActive);
+            await LoadScenes(new List<string> { scene});
         }
 
         /// <summary>
-        /// Load multiple scenes by a list of scene name
+        /// Load multiple scenes by a list of sceneToUnload name
         /// </summary>
         /// <param name="scenes"></param>
         /// <returns></returns>
-        public async UniTask LoadScenes(List<string> scenes, bool makeSceneActive = false)
+        public async UniTask LoadScenes(List<string> scenes)
         {
-            await LoadScenes(scenes, LoadSceneMode.Additive,makeSceneActive);
+            await LoadScenes(scenes, LoadSceneMode.Additive);
         }
 
         /// <summary>
-        /// Load a single scene by scene name without loading screen
+        /// Load a single sceneToUnload by sceneToUnload name without loading screen
         /// </summary>
         /// <param name="scene"></param>
         /// <returns></returns>
-        public async UniTask LoadSceneWithoutLoadingScreen(string scene, LoadSceneMode sceneMode = LoadSceneMode.Additive, bool makeSceneActive = false)
+        public async UniTask LoadSceneWithoutLoadingScreen(string scene, LoadSceneMode sceneMode = LoadSceneMode.Additive)
         {
-            SceneLoader sceneLoader = new SceneLoader(new List<string> { scene }, sceneMode, makeSceneActive);
-            sceneLoader.OnSceneLoadingComplete.AddListener(sceneLoader.ActivateScenes);
-            await sceneLoader.LoadSceneAsync();
+            await LoadWithoutLoadingScreen(new List<string> { scene }, sceneMode);
+        }
+
+
+        /// <summary>
+        /// Load multiple scenes by sceneToUnload name without loading screen
+        /// </summary>
+        /// <param name="scene"></param>
+        /// <returns></returns>
+        public async UniTask LoadScenesWithoutLoadingScreen(List<string> scene, LoadSceneMode sceneMode = LoadSceneMode.Additive)
+        {
+            await LoadWithoutLoadingScreen(scene, sceneMode);
         }
 
         /// <summary>
-        /// Unload a scene by scene name
+        /// Unload a sceneToUnload by sceneToUnload name
         /// </summary>
         /// <param name="scene"></param>
         /// <returns></returns>
@@ -73,7 +82,22 @@ namespace Studio23.SS2.SceneLoadingSystem.Core
         }
 
         /// <summary>
-        /// Unload all scene from build settings
+        /// Unload multiple by sceneToUnload name
+        /// </summary>
+        /// <param name="sceneToUnload"></param>
+        /// <returns></returns>
+        public async UniTask UnloadScenes(List<string> sceneToUnload)
+        {
+            List<UniTask> sceneToUnloadsTask = new List<UniTask>();
+            foreach (var scene in sceneToUnload)
+            {
+                sceneToUnloadsTask.Add(SceneLoader.UnloadScene(scene));
+            }
+            await UniTask.WhenAll(sceneToUnloadsTask);
+        }
+
+        /// <summary>
+        /// Unload all sceneToUnload from build settings
         /// </summary>
         public void UnloadAll()
         {
@@ -84,11 +108,18 @@ namespace Studio23.SS2.SceneLoadingSystem.Core
             }
         }
 
-        private async UniTask LoadScenes(List<string> scenes,LoadSceneMode sceneMode, bool makeSceneActive = false)
+        private async UniTask LoadWithoutLoadingScreen(List<string> scenes, LoadSceneMode sceneMode)
+        {
+            SceneLoader sceneLoader = new SceneLoader(scenes, sceneMode);
+            sceneLoader.OnSceneLoadingComplete.AddListener(sceneLoader.ActivateScenes);
+            await sceneLoader.LoadSceneAsync();
+        }
+
+        private async UniTask LoadScenes(List<string> scenes,LoadSceneMode sceneMode)
         {
             AbstractLoadingScreenUI loadingScreen= Instantiate(_loadingScreenPrefab).GetComponent<AbstractLoadingScreenUI>();
             loadingScreen.Initialize();
-            SceneLoader sceneLoader= new SceneLoader(scenes,sceneMode, makeSceneActive);
+            SceneLoader sceneLoader= new SceneLoader(scenes,sceneMode);
             sceneLoader.OnSceneProgress.AddListener(loadingScreen.UpdateProgress);
             sceneLoader.OnSceneLoadingComplete.AddListener(loadingScreen.OnLoadingDone);
             loadingScreen.OnValidAnyKeyPressEvent.AddListener(sceneLoader.ActivateScenes);
@@ -96,9 +127,15 @@ namespace Studio23.SS2.SceneLoadingSystem.Core
             await sceneLoader.LoadSceneAsync();
         }
 
-        public void SetActiveScene(string sceneName)
+        /// <summary>
+        /// Make a single scene active
+        /// </summary>
+        public async UniTask SetActiveScene(string activeScene)
         {
-            SceneManager.SetActiveScene(SceneManager.GetSceneByName(sceneName));
+            Scene sceneToActivate = SceneManager.GetSceneByName(activeScene);
+            if(!sceneToActivate.IsValid()) return;
+            await UniTask.WaitUntil(() => sceneToActivate.isLoaded);
+            SceneManager.SetActiveScene(sceneToActivate);
         }
     }
 }